# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "BFD",
    "contact": {}
  },
  "host": "localhost:8080",
  "basePath": "/api",
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/whplace/get": {
      "post": {
        "summary": "Get WH Places",
        "tags": [
          "WH Place"
        ],
        "operationId": "GetWHPlaces",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/GetWHPlacesRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/whplace": {
      "get": {
        "summary": "Get WH Place",
        "tags": [
          "WH Place"
        ],
        "operationId": "GetWHPlace",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/whplace/update": {
      "post": {
        "summary": "Update WH Place",
        "tags": [
          "WH Place"
        ],
        "operationId": "UpdateWHPlace",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateWHPlaceRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "GetWHPlacesRequest": {
      "title": "GetWHPlacesRequest",
      "example": {
        "search": "",
        "filters": {
          "lagar": [],
          "city": []
        },
        "organizer": ""
      },
      "type": "object",
      "properties": {
        "search": {
          "type": "string"
        },
        "filters": {
          "$ref": "#/definitions/Filters1"
        },
        "organizer": {
          "type": "string"
        }
      },
      "required": [
        "search",
        "filters",
        "organizer"
      ]
    },
    "Filters1": {
      "title": "Filters1",
      "example": {
        "lagar": [],
        "city": []
      },
      "type": "object",
      "properties": {
        "lagar": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "city": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "required": [
        "lagar",
        "city"
      ]
    },
    "UpdateWHPlaceRequest": {
      "title": "UpdateWHPlaceRequest",
      "example": {
        "id": "",
        "city": "",
        "rackSelfNumber": [],
        "rackNumber": "",
        "lagar": "",
        "organizer": ""
      },
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "rackSelfNumber": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "rackNumber": {
          "type": "string"
        },
        "lagar": {
          "type": "string"
        },
        "organizer": {
          "type": "string"
        }
      },
      "required": [
        "id",
        "city",
        "rackSelfNumber",
        "rackNumber",
        "lagar",
        "organizer"
      ]
    },
...
  },
...
}